requirements.txt
# ==============================================================================
# 文件: openmanus_minimal/requirements.txt
# 说明: 项目运行所需的核心 Python 依赖。
# 使用方法: pip install -r requirements.txt
# ==============================================================================

pydantic<2.0.0
python-dotenv==1.0.1
litellm==1.35.2
docker==7.0.0
toml==0.10.2